requirements.txt
torchvision
Pillow
opencv-python
numpy
requests